conversion

+++ src/event/source/windows.rs
@@ -39,7 +39,7 @@ impl EventSource for WindowsEventSource {
                         InputRecord::KeyEvent(record) => handle_key_event(record)?,
                         InputRecord::MouseEvent(record) => handle_mouse_event(record)?,
                         InputRecord::WindowBufferSizeEvent(record) => {
-                            Some(Event::Resize(record.size.x, record.size.y))
+                            Some(Event::Resize(record.size.x as u16, record.size.y  as u16))
                         }
                         _ => None,
                     };
